move: mproxy from repo to mf
