+++ api/asyncapi/mqtt.yml
@@ -6,7 +6,7 @@ info:
     url: https://www.apache.org/licenses/LICENSE-2.0
   version: '1.0.0'
   description: |
-    MQTT adapter provides an MQTT API for sending messages through the platform. MQTT adapter uses [mProxy](https://github.com/MainfluxLabs/mproxy) for proxying traffic between client and MQTT broker.
+    MQTT adapter provides an MQTT API for sending messages through the platform. MQTT adapter uses mProxy (`pkg/mproxy`) for proxying traffic between client and MQTT broker.
     Additionally, the MQTT adapter and the message broker are replicating the traffic between brokers.
 
 defaultContentType: application/json
