+++ cmd/mqtt/main.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/MainfluxLabs/mainflux"
 	authapi "github.com/MainfluxLabs/mainflux/auth/api/grpc"
+	"github.com/MainfluxLabs/mainflux/logger"
 	"github.com/MainfluxLabs/mainflux/mqtt"
 	mqttapi "github.com/MainfluxLabs/mainflux/mqtt/api"
 	mqttapihttp "github.com/MainfluxLabs/mainflux/mqtt/api/http"
@@ -28,14 +29,13 @@ import (
 	"github.com/MainfluxLabs/mainflux/pkg/messaging/brokers"
 	mqttpub "github.com/MainfluxLabs/mainflux/pkg/messaging/mqtt"
 	"github.com/MainfluxLabs/mainflux/pkg/messaging/nats"
+	mp "github.com/MainfluxLabs/mainflux/pkg/mproxy/mqtt"
+	"github.com/MainfluxLabs/mainflux/pkg/mproxy/session"
+	ws "github.com/MainfluxLabs/mainflux/pkg/mproxy/websocket"
 	"github.com/MainfluxLabs/mainflux/pkg/servers"
 	servershttp "github.com/MainfluxLabs/mainflux/pkg/servers/http"
 	"github.com/MainfluxLabs/mainflux/pkg/ulid"
 	thingsapi "github.com/MainfluxLabs/mainflux/things/api/grpc"
-	"github.com/MainfluxLabs/mproxy/logger"
-	mp "github.com/MainfluxLabs/mproxy/pkg/mqtt"
-	"github.com/MainfluxLabs/mproxy/pkg/session"
-	ws "github.com/MainfluxLabs/mproxy/pkg/websocket"
 	"github.com/cenkalti/backoff/v4"
 	kitprometheus "github.com/go-kit/kit/metrics/prometheus"
 	"github.com/go-redis/redis/v8"
